requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2==2.9.10
pygame==2.6.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.1
